build: Use git-version-gen
Idea behind this patch is to allow using git-version-gen script instead
of hand versioning.

git-version-gen is script which is able to use information from git
repository, file generated during make dist or gitattributes file
substituted during git-archive and pass it to autoconf as a main
version of package.

This simplifies releasing new version of package (only changing of fallback
version and making git tag is neede) and allows to build
non-released git snapshot with proper version number.

Similar method is used across all of the ClusterLabs projects so now it
is also in booth.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>